+++ Synologen.Core/Domain/Model/Orders/OptionalSequenceDefinition.cs
@@ -3,7 +3,7 @@ namespace Spinit.Wpc.Synologen.Core.Domain.Model.Orders
 	public class OptionalSequenceDefinition
 	{
 		public OptionalSequenceDefinition() { }
-		public OptionalSequenceDefinition(float? min, float? max, float? increment, bool disable)
+		public OptionalSequenceDefinition(decimal? min, decimal? max, decimal? increment, bool disable)
 		{
     		Min = min;
     		Max = max;
@@ -11,8 +11,8 @@ namespace Spinit.Wpc.Synologen.Core.Domain.Model.Orders
 			DisableDefinition = disable;
 		}
 		public virtual bool DisableDefinition { get; set; }
-        public virtual float? Min { get; set; }
-        public virtual float? Max { get; set; }
-        public virtual float? Increment { get; set; }
+        public virtual decimal? Min { get; set; }
+        public virtual decimal? Max { get; set; }
+        public virtual decimal? Increment { get; set; }
 	}
 }
